FileTrans_v1.1

+++ LanFileTransfer/src/com/wsjc/connection/FileRecv.java
@@ -1,23 +1,27 @@
 package com.wsjc.connection;
 
+import java.net.InetAddress;
 import java.net.ServerSocket;
 import java.net.Socket;
 import java.io.*;
 
 import com.wsjc.data.Data;
+import com.wsjc.tools.BasicThread;
 import com.wsjc.tools.ThreadMgr;
 import com.wsjc.view.FTView;
 
-public class FileRecv extends Thread {
+public class FileRecv extends BasicThread {
 
 	private ServerSocket server = null;
-	private Socket socket = null;
 	private FileOutputStream fos = null;
 	private BufferedInputStream bis = null;
-	private int port;
 	private FTView view = (FTView) ThreadMgr.getThread(ThreadMgr.FTVIEW);
 	private File file = null;
+	private InetAddress ip = null;
 	private long totalLen = 0;
+	private boolean isShutdown = false;
+	private Socket socket;
+	private int port;
 
 	public FileRecv(File file, long size, int port) {
 		totalLen = size;
@@ -26,22 +30,24 @@ public class FileRecv extends Thread {
 	}
 
 	public void shutdown() {
-		try {
-			if (fos != null) {
-				fos.close();
-				if (file.exists()) {
-					file.delete();
+		if (!isShutdown)
+			try {
+				if (fos != null) {
+					fos.close();
+					if (file.exists()) {
+						file.delete();
+					}
 				}
+				if (socket != null) {
+					socket.shutdownInput();
+					socket.close();
+				}
+				if (bis != null)
+					bis.close();
+				isShutdown = true;
+			} catch (Exception e) {
+				e.printStackTrace();
 			}
-			if (bis != null)
-				bis.close();
-			if (socket != null) {
-				socket.close();
-			}
-			if (server != null)
-				server.close();
-		} catch (Exception e) {
-		}
 	}
 
 	@Override
@@ -51,39 +57,43 @@ public class FileRecv extends Thread {
 		byte[] buf = null;
 		String output = "正在接收：" + file.getName() + " " + "已完成";
 		try {
-			view.appendText("等待接收："+file.getName());
+			view.appendText("准备接收："+file.getAbsolutePath());
 			server = new ServerSocket(port);
 			socket = server.accept();
-			view.appendText("正在接收："+file.getName()+"...");
-			bis = new BufferedInputStream(socket.getInputStream());
-			fos = new FileOutputStream(file);
-			buf = new byte[2048];
-			while ((read = bis.read(buf)) != -1 && length < totalLen) {
-				fos.write(buf, 0, read);
-				length += read;
-				per = (int) (length * 100 / totalLen);
-				if (per != tper) {
-					tper = per;
-					output += per + "%";
-					view.updateProgress(file, output);
+			if (socket != null) {
+				view.appendText("正在接收：" + file.getName() + "...");
+				bis = new BufferedInputStream(socket.getInputStream());
+				fos = new FileOutputStream(file);
+				buf = new byte[2048];
+				while ((read = bis.read(buf)) != -1 && length < totalLen) {
+					fos.write(buf, 0, read);
+					length += read;
+					per = (int) (length * 100 / totalLen);
+					if (per != tper) {
+						tper = per;
+						output += per + "%";
+						view.updateProgress(file, output);
+					}
+					fos.flush();
 				}
-				fos.flush();
 			}
 			if (length == totalLen) {
-				socket.shutdownInput();
-				view.appendText(java.util.Calendar.getInstance().toString()+"："+"传输完成："+file.getName());
+				view.appendText(java.util.Calendar.getInstance().toString() + "：" + "传输完成：" + file.getName());
 				view.updateProgress(file, file.getAbsolutePath());
-			}
-			else {
-				view.appendText("文件传输出错，请重试！"+file.getName());
+			} else {
+				view.appendText("文件传输出错，请重试！" + file.getName());
+				SendDataPkg send = new SendDataPkg();
+				send.sendPacket(new Data(Data.ERROR, "FileSend" + file.getName()), ip);
 			}
 			shutdown();
+			ThreadMgr.remove("FileRecv" + file.getName());
 		} catch (IOException e) {
-			if(length<totalLen) {
-				Data d = new Data(Data.ERROR);
-				SendDataPkg send = new SendDataPkg(d, socket.getInetAddress());
-				send.sendPacket();
+			if (length < totalLen) {
+				SendDataPkg send = new SendDataPkg();
+				send.sendPacket(new Data(Data.ERROR, "FileSend" + file.getName()), ip);
 			}
+			shutdown();
+			ThreadMgr.remove("FileRecv" + file.getName());
 			e.printStackTrace();
 		}
 	}
